+++ scripts/macros/make-secret-group-test.js
@@ -1,4 +1,17 @@
-let group = game.gmtoolkit.utility.getGroup("nonparty")
+// === Set up group 
+const groupType = "party";  // choose from, eg, "party", "tokens", "adversaries", "entourage", "npcTokens", etc
+const present = true; // choose from (to only include actors in the viewed scene): undefined, true or false 
+const interaction = undefined;  // choose from: undefined, "selected", "targeted"
+
+const group = game.gmtoolkit.utility.getGroup(groupType, {present, interaction}).filter(
+    g => (
+        g.documentName !== "User" && 
+        g.actor?.type !== "vehicle" && 
+        g.type !== "vehicle"
+        )
+    );
+// Exit with error if no actors in group
+if (!group.length) return ui.notifications.error(game.i18n.localize("GMTOOLKIT.Message.MakeSecretGroupTest.NoGroup"))
 
 // === Set up test parameters
 let targetSkill = "Perception";  // eg, basic skill
@@ -18,40 +31,41 @@ let testModifier = +20 // game.wfrp4e.config.difficultyModifiers[("average")]
 // --- accepted difficultyModifier values are listed in the REFERENCES section below
 
 // === Carry out the test if you are a GM
-if (game.user.isUniqueGM) 
-    {makeSecretPartyTest(targetSkill)} 
+if (game.user.isGM) 
+    {makeSecretGroupTest(targetSkill)} 
 else 
     {ui.notifications.error(game.i18n.localize("GMTOOLKIT.Message.MakeSecretPartyTest.NoPermission"), {})
 };
 
 // === Where the magic happens. 
 
-async function makeSecretPartyTest(targetSkill) {
-    // let party = Array.from(game.actors).filter(a => a.hasPlayerOwner && a.type == "character");
-    for (let pc of group) {
+async function makeSecretGroupTest(targetSkill) {
+    for (let member of group) {
+        // make sure to get the actor
+        let actor = member.actor || member;
         // stack tests rather than await: cancelling a non-silent test during an await(ed) call will abandon all subsequent tests
-        let pcSkill = game.gmtoolkit.utility.hasSkill(pc, targetSkill, "silent")
-        if (pcSkill != null) { 
+        let actorSkill = game.gmtoolkit.utility.hasSkill(actor, targetSkill, "silent")
+        if (actorSkill !== undefined) { 
             // Roll against the skill if the actor has it ...
-            pc.setupSkill(pcSkill, {bypass: silentTest, testModifier, slBonus, successBonus, rollMode, title : `${pcSkill.name} Test (${pc.name})`}).then(setupData => {pc.basicTest(setupData)});
+            actor.setupSkill(actorSkill, {bypass: silentTest, testModifier, slBonus, successBonus, rollMode, title : `${actorSkill.name} Test (${actor.name})`}).then(setupData => {actor.basicTest(setupData)});
         } else {
             // ... or fallback to underlying characteristic if they don't
-            pcSkill = await game.wfrp4e.utility.findSkill(targetSkill)
+            actorSkill = await game.wfrp4e.utility.findSkill(targetSkill)
             // TODO: optionally step-adjust the difficulty in case of fallback
-            if (pcSkill.advanced.value == "adv" && !game.settings.get("wfrp4e-gm-toolkit", "fallbackAdvancedSkills")) {
-                ui.notifications.info(`${game.i18n.format("GMTOOLKIT.Message.MakeSecretPartyTest.AbortAdvancedSkillTest", {character: pc.name, skill: targetSkill})}`)
+            if (actorSkill.advanced.value === "adv" && !game.settings.get("wfrp4e-gm-toolkit", "fallbackAdvancedSkills")) {
+                ui.notifications.info(`${game.i18n.format("GMTOOLKIT.Message.MakeSecretPartyTest.AbortAdvancedSkillTest", {character: actor.name, skill: targetSkill})}`)
                 continue 
             }
-            let skillCharacteristic = game.wfrp4e.config.characteristics[pcSkill.characteristic.value]
-            pc.setupCharacteristic(pcSkill.characteristic.value, {bypass: silentTest, testModifier, slBonus, successBonus, rollMode, title : `${skillCharacteristic} Test for ${pcSkill.name} (${pc.name})`}).then(setupData => {pc.basicTest(setupData)});
+            let skillCharacteristic = game.wfrp4e.config.characteristics[actorSkill.characteristic.value]
+            actor.setupCharacteristic(actorSkill.characteristic.value, {bypass: silentTest, testModifier, slBonus, successBonus, rollMode, title : `${skillCharacteristic} Test for ${actorSkill.name} (${actor.name})`}).then(setupData => {actor.basicTest(setupData)});
         }        
     }
 }
 
 /* ==========
  * MACRO: Make Secret Group Test
- * VERSION: 0.9.3
- * UPDATED: 2022-05-23
+ * VERSION: 0.9.4
+ * UPDATED: 2022-07-06
  * DESCRIPTION: Template for GMs to run secret skill tests for character groups. 
  * TIP: Create copies of this macro and adjust the parameters (and title and/or icon) for different skill tests. Add them to your macro bar for one-click resolution. 
  * TIP: By default, tests are rolled blind. Right-click a test result in the chat log to show the results to players.