+++ CHANGELOG.md
@@ -42,6 +42,10 @@ See [Issue Backlog](../../issues) and [Roadmap](../../milestones).
 - *Changed* various references to use localization keys for skills, talents and status provided by WFRP4e system. [[#110](https://github.com/Jagusti/fvtt-wfrp4e-gmtoolkit/issues/110)]
   - Feature: Token Hud Extension
   - Macros: Send Dark Whispers, Reset Fortune, Set Token Vision and Light
+- *Changed* Make Secret Party Test to **Make Secret Group Test**, using group selection functionality. 
+  - GMs can now choose to roll tests for any non-User group type (including "party", "tokens", "adversaries", with or without active, present and interaction filters). 
+  - Vehicles are filtered out of group constituents. 
+- *Fixed* issue where Assistant GMs could not make secret (party) group tests.
 
 ## [Version 0.9.3](https://github.com/Jagusti/fvtt-wfrp4e-gmtoolkit/releases/tag/v0.9.3)  (2022-05-25)
 - *Fixed* missing Token Hud Extension options for players who don't have access to configure tokens. The layout of Token Hud Extensions has been reorganised as a result of this change. [[#67](https://github.com/Jagusti/fvtt-wfrp4e-gmtoolkit/issues/67)]